Prevent calling make_optional<int&>(i) or make_optional<const int&>(i)
The status quo in C++23 is that
    int i;
    auto o = make_optional<int&>(i);
produces an `optional<int>`, not an `optional<int&>`. This would be horribly
confusing to propagate into C++26 — users would naturally expect to receive
an `optional<int&>`. Worse,
    auto o = make_optional<int&>(std::ref(i));
actually *would* give you an `optional<int&>`, because it would use a different
overload of `make_optional`! So, we propose to add a "Mandates" element to
`make_optional` that requires its explicit template argument (if any) to be
a non-array object type.
This breaks the "status quo" example at the top of this comment (requiring
an Annex C entry); but it successfully prevents the pitfalls.

Thanks to Tomasz Kamiński for the template-programming trick, which matches
the technique used in https://eel.is/c++draft/out.ptr .